refactor: remove tree size limit
BREAKING CHANGE: As we are creating optimised trees (see https://github.com/snyk/nodejs-lockfile-parser/pull/93), and that in most places we don't serialise the DepTree object, we don't need to really care about the size of the tree, but on the number of dependencies.
After reviewing most of the "big trees" tickets in Zendesk, the number of dependencies are limited by thousands, which shouldn't be a problem any more.

This removes the limit of the tree size.

This is a breaking change, as we must verify that this plugin consumers doesn't serialise this object.