feat: allow deeply out of sync lockfile
Yarn workspaces introduces complexity where local packages
are mentioned as required dependency but are not present on the lockfile
this means that the lockfile is considered deeply out of sync by the plugin.

Allow to skip the check if user is calling with StrictOutOfSync=false not just
for top level dependencies but also the transitives as a package can rely on another local workspace which lists a workspace as a dep.